fixed mobile popup and added package for fix lib error, alos I added some meta tags to pages
