+++ app/signup/page.tsx
@@ -195,7 +195,7 @@ function Signup() {
       <p className="text-center text-sm text-gray-500 mt-2 font-normal">
         Already have an account?{" "}
         <Link
-          href="/oldPages/login"
+          href="/login"
           className="text-my-orange hover:cursor-pointer font-bold"
         >
           Login
